Check a flag file to decide whether to boot into recovery
With these changes, u-boot checks for the existence of
/persist/reboot-recovery to decide when to reboot into recovery (in
addition to holding the volume up key). By including a blank persist
partition in the device tarball, we ensure that the flag file is
deleted when an update completes. Otherwise, we can delete the flag
file early in Jumpdrive's init.

Additionally, disable the framebuffer console on the internal screen
and set 'systempart' and 'devnum' kernel cmdline arguments using
u-boot's devnum variable. This will help us run on either eMMC or SD.

This isn't a good option, but it will do until we find something else.