Set voice calls to 8kHz
Pulse defaults to 44.1kHz on all audio sinks. Most of the time this is
fine, but during audio calls on the PinePhone it gets hairy. The
hardware audio routing can't handle call audio at 8kHz while Pulse also
drives it at a different sample rate.

By forcing an 8k playback rate in the VoiceCall profile, we avoid this
problem. Any notifications or alarms that play during the call sound
8kHz-y, but that's better than the call audio failing.