Use an older, but working, u-boot
Unfortunately, commit c77d2a43 in crust causes problems with Ubuntu
Touch with the 5.6 kernel. It does not cause problems with the 5.10
kernel, but we're not ready to upgrade yet.

By using a u-boot build from before that commit, we avoid the problem.

Fixes https://gitlab.com/ubports/community-ports/pinephone/-/issues/259

https://github.com/crust-firmware/crust/commit/c77d2a43e9e32d2333bd6454ea931f64d6b20d36