Modern modem problems require modern modem solutions
Problem: The modem gets issued a USB reset on wake. When this happens,
oFono gets confused and thinks a brand new modem was connected.
Alternatively, oFono gets more confused and never recognizes the new
or old modem, basically never working again.

Solution: Restart oFono and the modem when a USB reset is detected.

Currently, this is done by starting a new 'cellular-reset' Upstart job
via a udev rule whenever the modem's AT interface is detected. This
turns off the modem and turns it back on again while oFono does the
same.

Unfortunately it's not possible to just restart oFono and not the modem.
After the USB reset, oFono refuses to probe the modem until it's been
power-cycled. It doesn't matter how many times oFono is restarted. The
enable-modem operation times out and the modem remains disabled.

This also moves the logic from enable-eg25-ofono back into the ofono job
since having it removed was creating more problems with this job
dependency tree.