Fix random boot hang
The PinePhone randomly hung on boot when using this wip device image
because gpsd was failing to start. gpsd was failing to start because the
hardware it wanted to use wasn't available.

The problem was actually a bit more interesting, and caused partially by
system-watchdog. gpsd would quickly exit with status 1 a bunch of times
and hit its respawn limit. That caused system-watchdog to bite and
attempt to reboot the system. For some reason, the reboot never
completed. At that point, the system was stuck in a zombie state where
it wouldn't start or stop any services, and just hung at the Pine64 logo