Add PineTab ucm file
This ucm file was generated by taking [1], adding a 'cdev' at the
beginning of each sequence, adding '.0' to the end of each SectionDevice
header, and replacing all instances of '${CardId},0' with
'CARD=PineTab'. That's an effective conversion to ucm1.

Unfortunately, headphone jack detection doesn't work on kernel 5.13 as
released right now. But even so, I put the JackControl in place.

[1] https://gitlab.com/postmarketOS/pmaports/-/blob/master/device/community/device-pine64-pinetab/ucm/HiFi.conf